requirements.txt
Django==5.2.*
python-dotenv==1.1.*
psycopg==3.2.*
psycopg[c]==3.2.*
Pillow==11.1.*
djangorestframework==3.16.*
channels==4.2.*
channels-redis==4.2.*
daphne==4.1.*